Create MemoryManager class to handle shared memory
