+++ src/c++/perf_analyzer/CMakeLists.txt
@@ -1,4 +1,4 @@
-# Copyright 2020-2022, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# Copyright 2020-2023, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
 # modification, are permitted provided that the following conditions
@@ -51,6 +51,7 @@ set(
   report_writer.cc
   mpi_utils.cc
   metrics_manager.cc
+  memory_manager.cc
 )
 
 set(
@@ -76,6 +77,7 @@ set(
   metrics.h
   metrics_manager.h
   test_load_manager_base.h
+  memory_manager.h
 )
 
 add_executable(
@@ -161,10 +163,11 @@ add_executable(
   test_metrics_manager.cc
   test_perf_utils.cc
   test_report_writer.cc
-  test_request_rate_manager.cc
   client_backend/triton/test_triton_client_backend.cc
-  test_concurrency_manager.cc
-  test_custom_load_manager.cc
+  mock_memory_manager.h
+  # test_request_rate_manager.cc
+  # test_concurrency_manager.cc
+  # test_custom_load_manager.cc
   $<TARGET_OBJECTS:json-utils-library>
 )
