+++ src/c++/perf_analyzer/concurrency_worker.h
@@ -1,4 +1,4 @@
-// Copyright 2022, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+// Copyright 2022 - 2023, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 //
 // Redistribution and use in source and binary forms, with or without
 // modification, are permitted provided that the following conditions
@@ -78,13 +78,14 @@ class ConcurrencyWorker : public LoadWorker {
       std::unordered_map<std::string, SharedMemoryData>& shared_memory_regions,
       std::condition_variable& wake_signal, std::mutex& wake_mutex,
       size_t& active_threads, bool& execute, std::atomic<uint64_t>& curr_seq_id,
-      std::uniform_int_distribution<uint64_t>& distribution)
+      std::uniform_int_distribution<uint64_t>& distribution,
+      const std::shared_ptr<MemoryManager> memory_manager)
       : LoadWorker(
             thread_stat, parser, data_loader, factory, sequence_stat,
             shared_memory_regions, backend_kind, shared_memory_type,
             on_sequence_model, async, streaming, batch_size, using_json_data,
             sequence_length, start_sequence_id, sequence_id_range, curr_seq_id,
-            distribution, wake_signal, wake_mutex, execute),
+            distribution, wake_signal, wake_mutex, execute, memory_manager),
         thread_config_(thread_config), max_concurrency_(max_concurrency),
         threads_config_(threads_config), active_threads_(active_threads)
   {
