+++ src/c++/perf_analyzer/request_rate_manager.cc
@@ -1,4 +1,4 @@
-// Copyright 2020-2022, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+// Copyright 2020-2023, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 //
 // Redistribution and use in source and binary forms, with or without
 // modification, are permitted provided that the following conditions
@@ -180,7 +180,7 @@ RequestRateManager::MakeWorker(
       on_sequence_model_, async_, max_threads_, using_json_data_, streaming_,
       shared_memory_type_, batch_size_, sequence_stat_, shared_memory_regions_,
       wake_signal_, wake_mutex_, execute_, curr_seq_id_, start_time_, schedule_,
-      gen_duration_, distribution_);
+      gen_duration_, distribution_, memory_manager_);
 }
